requirements.txt
pywhatkit
pyautogui
pynput
public_ip 
telebot 
telethon 
python-dotenv
